+++ src/main/scala/millfork/compiler/mos/MosMacroExpander.scala
@@ -34,7 +34,7 @@ object MosMacroExpander extends MacroExpander[AssemblyLine] {
           case x => x
         }
       case (AssemblyOrMacroParam(typ, Placeholder(ph, phType), AssemblyParameterPassingBehaviour.ByConstant), actualParam) =>
-        ctx.env.eval(actualParam).getOrElse(ctx.env.errorConstant("Non-constant expression was passed to an inlineable function as a `const` parameter", actualParam.position))
+        ctx.env.eval(actualParam).getOrElse(ctx.env.errorConstant("Non-constant expression was passed to an inlineable function as a `const` parameter", Some(actualParam), actualParam.position))
         actualCode = actualCode.map {
           case a@MosAssemblyStatement(_, _, expr, _) =>
             a.copy(expression = expr.replaceVariable(ph, actualParam))