+++ src/main/scala/millfork/compiler/z80/Z80MacroExpander.scala
@@ -36,7 +36,7 @@ object Z80MacroExpander extends MacroExpander[ZLine] {
             case x => x
           }
         case (AssemblyOrMacroParam(typ, Placeholder(ph, phType), AssemblyParameterPassingBehaviour.ByConstant), actualParam) =>
-          ctx.env.eval(actualParam).getOrElse(ctx.env.errorConstant("Non-constant expression was passed to an inlineable function as a `const` parameter", actualParam.position))
+          ctx.env.eval(actualParam).getOrElse(ctx.env.errorConstant("Non-constant expression was passed to an inlineable function as a `const` parameter", Some(actualParam), actualParam.position))
           actualCode = actualCode.map {
             case a@Z80AssemblyStatement(_, _, _, expr, _) =>
               a.copy(expression = expr.replaceVariable(ph, actualParam))